+++ CHANGELOG.md
@@ -6,10 +6,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-* Create a person;
 * Update a person;
 * Search a person;
 * List persons;
-* Create person's address;
 * List person's address;
 * Set main person's main address;
+
+## [0.1.0]() - 2023-01-12
+### Added
+* H2 database;
+* Person model/table;
+* Address model/table;
+
+## [0.2.0]() - 2023-01-15
+### Added
+* Create a person;
+* Create person's address;